Add mei-friend
This PR adds a link to mei-friend, a browser-based graphical music encoding editor. It is capable of opening encodings in a variety of formats, but converts them to MEI, the format of the Music Encoding Initiative for editing. The editor supports a number of basic and advanced features, documented at https://mei-friend.github.io/. The browser is open-source and community developed, and widely used by members of the MEI, counting typically more than 500 unique user sessions a month at its official instance hosted by mdw - University of Music and Performing Arts Vienna.